fix gulp and move utils
