+++ src/components/Select/Select/index.js
@@ -1,9 +1,9 @@
 import cx from 'classnames'
 import invariant from 'invariant'
 import PropTypes from 'prop-types'
-import React, {Component} from 'react'
-import dataFilter from '../../../utils/dataFilter'
-import shouldComponentUpdate from '../../../utils/shouldComponentUpdate'
+import React, { Component } from 'react'
+import dataFilter from '../../_utils/dataFilter'
+import shouldComponentUpdate from '../../_utils/shouldComponentUpdate'
 import ClearableInput from '../../ClearableInput'
 import SelectDropdown from '../../SelectDropdown'
 import TextOverflow from '../../TextOverflow'
