+++ src/components/dialog/Dialog.js
@@ -2,8 +2,8 @@ import classlist from 'classlist'
 import PropTypes from 'prop-types'
 import React from 'react'
 import ReactDOM from 'react-dom'
-import EventPool from '../../utils/EventPool'
-import ToggleNode from '../../utils/ToggleNode'
+import EventPool from '../_utils/EventPool'
+import ToggleNode from '../_utils/ToggleNode'
 import DialogContainer from './DialogContainer'
 import './index.less'
